fixing code-climate issues
